+++ WorldQuests.lua
@@ -733,6 +733,10 @@ local RetrieveWorldQuests = function(mapId)
 									rewardType[#rewardType+1] = CONSTANTS.REWARD_TYPES.THE_HARATI
 									quest.reward.TheHaratiAmount = currency.amount
 									if BWQ:C("showTheHarati") then quest.hide = false end		
+								elseif currencyId == 3310 then -- Coffer Key Shards
+									rewardType[#rewardType+1] = CONSTANTS.REWARD_TYPES.COFFER_KEY_SHARDS
+									quest.reward.CofferKeyShardsAmount = currency.amount
+									if BWQ:C("showCofferKeyShards") then quest.hide = false end		
 								else 
 									if BWQcfg.spewDebugInfo then print(string.format("[BWQ] Unhandled currency: ID %s", currencyId)) end
 								end
@@ -1006,6 +1010,8 @@ local RetrieveWorldQuests = function(mapId)
 									BWQ.totalTheSingularity = BWQ.totalTheSingularity + quest.reward.TheSingularityAmount
 								elseif rtype == CONSTANTS.REWARD_TYPES.THE_HARATI then
 									BWQ.totalTheHarati = BWQ.totalTheHarati + quest.reward.TheHaratiAmount
+								elseif rtype == CONSTANTS.REWARD_TYPES.COFFER_KEY_SHARDS then
+									BWQ.totalCofferKeyShards = BWQ.totalCofferKeyShards + quest.reward.CofferKeyShardsAmount
 								end
 							end
 						end
@@ -1206,7 +1212,7 @@ function BWQ:UpdateQuestData()
 	BWQ.totalPolishedPetCharms, BWQ.totalCouncilofDornogal, BWQ.totalTheWeaver, BWQ.totalTheGeneral, BWQ.totalTheVizier = 0, 0, 0, 0, 0
 	BWQ.totalXP, BWQ.totalBronzeCelebrationToken, BWQ.totalWeatheredUndermineCrest, BWQ.totalCarvedUndermineCrest, BWQ.totalTheCartelsOfUndermine = 0, 0, 0, 0, 0
 	BWQ.totalTheBilgewaterCartel, BWQ.totalTheBlackwaterCartel, BWQ.totalTheSteamwheedleCartel, BWQ.totalTheVentureCompany, BWQ.totalWeatheredEtherealCrest = 0, 0, 0, 0, 0
-	BWQ.totalVoidlightMarl, BWQ.totalTheAmaniTribe, BWQ.totalTheSingularity, BWQ.totalTheHarati = 0, 0, 0, 0
+	BWQ.totalVoidlightMarl, BWQ.totalTheAmaniTribe, BWQ.totalTheSingularity, BWQ.totalTheHarati, BWQ.totalCofferKeyShards = 0, 0, 0, 0, 0
 	BWQ.totalTwilightsBladeInsignia = 0
 
 	for mapId in next, BWQ.MAP_ZONES[BWQ.expansion] do
@@ -1834,6 +1840,7 @@ function BWQ:UpdateBlock()
 		if BWQ:C("brokerShowTheAmaniTribe") 		and BWQ.totalTheAmaniTribe > 0			then brokerString = string.format("%s|T7505698:16:16|t %d  ", brokerString, BWQ.totalTheAmaniTribe) end
 		if BWQ:C("brokerShowTheSingularity") 		and BWQ.totalTheSingularity > 0 		then brokerString = string.format("%s|T7505702:16:16|t %d  ", brokerString, BWQ.totalTheSingularity) end
 		if BWQ:C("brokerShowTheHarati") 			and BWQ.totalTheHarati > 0 				then brokerString = string.format("%s|T7505704:16:16|t %d  ", brokerString, BWQ.totalTheHarati) end
+		if BWQ:C("brokerShowCofferKeyShards") 		and BWQ.totalCofferKeyShards > 0		then brokerString = string.format("%s|TInterface\\Icons\\inv_gizmo_hardenedadamantitetube:16:16|t %d  ", brokerString, BWQ.totalCofferKeyShards) end
 
 		-- If necessary, the following command can be used to get the icon IDs above (3354 is an example of the factionID, as found in Constants.lua)
 		-- > /dump C_CurrencyInfo.GetCurrencyInfo(3354).iconFileID